Fix -- Add .readthedocs.yml
Problem:
--------

Currently documentation is not showing code snippets in the API
Reference section. Looking at the build logs, there is these warnings:

> WARNING: autodoc: failed to import function 'fake_word' from module
> 'anon'; the following exception was raised:
> No module named 'chunkator'

ReadTheDocs does not performs `pip install .` by default, which means
that when trying to `import anon` it will fail due to dependencies not
being installed.

Solution
--------

Add the recommended `.readthedocs.yml` configuration file as it allows
specifying ReadTheDocs to run `pip install .`, which will install
dependencies and let code snippets to be processed and appear in the
documentation